cleaning

+++ packages/cli-e2e/entrypoints/entry.js
@@ -49,17 +49,20 @@ const isImagePresent = () => {
 if (!isImagePresent()) {
   console.log('Building docker image');
   execSync(`docker build -t ${DOCKER_IMAGE_NAME} ${dockerFilePath}`, {
-    stdio: ['ignore', 'inherit', 'inherit'],
+    stdio: 'ignore',
   });
 }
 try {
   execSync(
-    `docker run --name=${DOCKER_CONTAINER_NAME} -v "${repoHostPath}:${repoDockerPath}" -p "9229:9229" -t --cap-add=SYS_ADMIN ${DOCKER_IMAGE_NAME} ${dockerEntryPoint}`,
-    // `docker run --name=${DOCKER_CONTAINER_NAME} -v "${repoHostPath}:${repoDockerPath}" -it --cap-add=SYS_ADMIN ${DOCKER_IMAGE_NAME} /bin/bash`,
+    `docker run --name=${DOCKER_CONTAINER_NAME} -v "${repoHostPath}:${repoDockerPath}" -p "9229:9229" -${
+      process.argv[2] === '--bash' ? 'i' : 'it'
+    } --cap-add=SYS_ADMIN ${DOCKER_IMAGE_NAME} ${dockerEntryPoint}`,
     {stdio: ['ignore', 'inherit', 'inherit']}
   );
 } finally {
-  execSync(`docker container rm ${DOCKER_CONTAINER_NAME}`, {
-    stdio: ['ignore', 'inherit', 'inherit'],
-  });
+  if (!process.env.CI) {
+    execSync(`docker container rm ${DOCKER_CONTAINER_NAME}`, {
+      stdio: 'ignore',
+    });
+  }
 }
